# static/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Cognixus Todo List API",
		"description": "Todo List REST API",
		"version": "1.0.0",
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
	"tags": [
		{
			"name": "Todo API",
			"description": "REST API"
		}
	],
	"servers": [
		{
			"url": "/"
		}
	],
	"components": {
		"schemas": {
			"token": {
				"type": "string"
			},
			"id": {
				"type": "string",
				"properties": {
					"id": {
						"type": "string"
					}
				}
			},
			"signup": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"confirm_password": {
						"type": "string"
					}
				}
			},
			"login": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				}
			},
			"update": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string"
					},
					"tags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"completed": {
						"type": "boolean"
					}
				}
			},
			"add": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string"
					},
					"tags": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				}
			},
			"error": {
				"type": "object",
				"properties": {
					"Status": {
						"type": "integer"
					},
					"Error": {
						"type": "string"
					}
				}
			},
			"authRes": {
				"type": "object",
				"properties": {
					"Status": {
						"type": "integer"
					},
					"Message": {
						"type": "string"
					},
					"Auth": {
						"type": "object",
						"properties": {
							"token": {
								"type": "string"
							}
						}
					}
				}
			},
			"todoListRes": {
				"type": "object",
				"properties": {
					"Status": {
						"type": "integer"
					},
					"Message": {
						"type": "string"
					},
					"data": {
						"type": "array",
						"items": {
							"type": "object",
							"properties": {
								"_id": {
									"type": "string"
								},
								"completed": {
									"type": "boolean"
								},
								"date": {
									"type": "string"
								},
								"tags": {
									"type": "array",
									"items": {
										"type": "string"
									}
								},
								"title": {
									"type": "string"
								},
								"userID": {
									"type": "string"
								}
							}
						}
					}
				}
			},
			"todoRes": {
				"type": "object",
				"properties": {
					"Status": {
						"type": "integer"
					},
					"Message": {
						"type": "string"
					},
					"data": {
						"type": "object",
						"properties": {
							"_id": {
								"type": "string"
							},
							"completed": {
								"type": "boolean"
							},
							"date": {
								"type": "string"
							},
							"tags": {
								"type": "array",
								"items": {
									"type": "string"
								}
							},
							"title": {
								"type": "string"
							},
							"userID": {
								"type": "string"
							}
						}
					}
				}
			},
			"deleteTodoRes": {
				"type": "object",
				"properties": {
					"Status": {
						"type": "integer"
					},
					"Message": {
						"type": "string"
					}
				}
			}
		}
	},
	"paths": {
		"/signup": {
			"post": {
				"tags": [
					"Signup User Account"
				],
				"summary": "Signup User Account",
				"requestBody": {
					"description": "Signup User Account",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/signup"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "200 | SUCCESS",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/authRes"
								}
							}
						}
					},
					"401": {
						"description": "401 | FAILED",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/error"
								}
							}
						}
					}
				}
			}
		},
		"/login": {
			"post": {
				"tags": [
					"Login User Account"
				],
				"summary": "Login User Account",
				"requestBody": {
					"description": "Login User Account",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/login"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "200 | SUCCESS",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/authRes"
								}
							}
						}
					},
					"401": {
						"description": "401 | FAILED",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/error"
								}
							}
						}
					}
				}
			}
		},
		"/todolist": {
			"get": {
				"tags": [
					"Retrieve Todo List"
				],
				"summary": "Get All Todos",
				"parameters": [
					{
						"name": "token",
						"in": "path",
						"description": "Logged In Token",
						"required": true,
						"type": "string",
						"schema": {
							"$ref": "#/components/schemas/token"
						}
					}
				],
				"responses": {
					"200": {
						"description": "200 | SUCCESS",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/todoListRes"
								}
							}
						}
					},
					"401": {
						"description": "401 | FAILED",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/error"
								}
							}
						}
					}
				}
			}
		},
		"/todolist/{id}": {
			"get": {
				"tags": [
					"Get A Specific Todo"
				],
				"summary": "Get Todo",
				"parameters": [
					{
						"name": "token",
						"in": "path",
						"description": "Logged In Token",
						"required": true,
						"type": "string",
						"schema": {
							"$ref": "#/components/schemas/token"
						}
					},
					{
						"name": "id",
						"in": "path",
						"description": "todo id to get",
						"required": true,
						"type": "string",
						"schema": {
							"$ref": "#/components/schemas/id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "200 | SUCCESS",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/todoRes"
								}
							}
						}
					},
					"401": {
						"description": "401 | FAILED",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/error"
								}
							}
						}
					}
				}
			}
		},
		"/todolist/add": {
			"post": {
				"tags": [
					"Add Todo"
				],
				"summary": "Add Todo",
				"parameters": [
					{
						"name": "token",
						"in": "path",
						"description": "Logged In Token",
						"required": true,
						"type": "string",
						"schema": {
							"$ref": "#/components/schemas/token"
						}
					}
				],
				"requestBody": {
					"description": "Add Todo",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/add"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "200 | SUCCESS",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/todoRes"
								}
							}
						}
					},
					"401": {
						"description": "401 | FAILED",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/error"
								}
							}
						}
					}
				}
			}
		},
		"/todolist/update/{id}": {
			"put": {
				"tags": [
					"Update Todo"
				],
				"parameters": [
					{
						"name": "token",
						"in": "path",
						"description": "Logged In Token",
						"required": true,
						"type": "string",
						"schema": {
							"$ref": "#/components/schemas/token"
						}
					},
					{
						"name": "id",
						"in": "path",
						"description": "Todo ID to Update",
						"required": true,
						"type": "string",
						"schema": {
							"$ref": "#/components/schemas/id"
						}
					}
				],
				"summary": "Update Todo",
				"requestBody": {
					"description": "Update Existing todo On Database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/update"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "200 | SUCCESS",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/todoRes"
								}
							}
						}
					},
					"401": {
						"description": "401 | FAILED",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/error"
								}
							}
						}
					}
				}
			}
		},
		"/todolist/remove/{id}": {
			"delete": {
				"tags": [
					"Delete Todo"
				],
				"parameters": [
					{
						"name": "token",
						"in": "path",
						"description": "Logged In Token",
						"required": true,
						"type": "string",
						"schema": {
							"$ref": "#/components/schemas/token"
						}
					},
					{
						"name": "id",
						"in": "path",
						"description": "Todo ID to Delete",
						"required": true,
						"type": "string",
						"schema": {
							"$ref": "#/components/schemas/id"
						}
					}
				],
				"summary": "Delete Todo",
				"responses": {
					"200": {
						"description": "200 | SUCCESS",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/deleteTodoRes"
								}
							}
						}
					},
					"401": {
						"description": "401 | FAILED",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/error"
								}
							}
						}
					}
				}
			}
		}
	}
}